Move endpoints to dedicated controller
